@@ -1,6 +1,6 @@
 
 buildscript {
-	ext.kotlin_version = '1.6.21'
+	ext.kotlin_version = '1.8.20'
 
 	repositories {
 		google()
