@@ -522,7 +522,9 @@ public class NetworkModule extends KrollModule
 			KrollDict event = new KrollDict();
 			event.put("success", true);
 			event.put("type", "remote");
-			successCallback.callAsync(getKrollObject(), new KrollDict());
+			if (successCallback != null) {
+				successCallback.callAsync(getKrollObject(), new KrollDict());
+			}
 			return;
 		}
 
@@ -536,16 +538,23 @@ public class NetworkModule extends KrollModule
 				@NonNull TiBaseActivity activity, int requestCode,
 				@NonNull String[] permissions, @NonNull int[] grantResults)
 			{
-				Boolean isGranted = grantResults[0] == PackageManager.PERMISSION_GRANTED;
+				Boolean isGranted = false;
+				if (grantResults.length > 0) {
+					isGranted = grantResults[0] == PackageManager.PERMISSION_GRANTED;
+				}
 
 				KrollDict event = new KrollDict();
 				event.put("success", isGranted);
 				event.put("type", "remote");
 
 				if (isGranted) {
-					successCallback.callAsync(getKrollObject(), event);
+					if (successCallback != null) {
+						successCallback.callAsync(getKrollObject(), event);
+					}
 				} else {
-					errorCallback.callAsync(getKrollObject(), event);
+					if (errorCallback != null) {
+						errorCallback.callAsync(getKrollObject(), event);
+					}
 				}
 
 				// Unregister this callback.
