@@ -22,6 +22,7 @@ import android.graphics.drawable.Drawable;
 import android.graphics.drawable.ShapeDrawable;
 import android.graphics.drawable.shapes.RectShape;
 import android.os.Handler;
+import android.os.Parcelable;
 import android.os.SystemClock;
 import android.view.MotionEvent;
 import android.view.View;
@@ -342,7 +343,10 @@ public class TiTableView extends TiSwipeRefreshLayout implements OnSearchChangeL
 			final TableViewHolder firstVisibleHolder =
 				(TableViewHolder) recyclerView.getChildViewHolder(firstVisibleView);
 			final TableViewRowProxy firstVisibleProxy = (TableViewRowProxy) firstVisibleHolder.getProxy();
-			final int firstVisibleIndex = firstVisibleProxy.getIndexInSection();
+			int firstVisibleIndex = -1;
+			if (firstVisibleProxy != null) {
+				firstVisibleIndex = firstVisibleProxy.getIndexInSection();
+			}
 			payload.put(TiC.PROPERTY_FIRST_VISIBLE_ITEM, firstVisibleIndex);
 		}
 
@@ -668,9 +672,11 @@ public class TiTableView extends TiSwipeRefreshLayout implements OnSearchChangeL
 
 					// Maintain true row index.
 					row.index = index++;
+					boolean alwaysInclude = row.getProperties()
+						.optBoolean(TiC.PROPERTY_FILTER_ALWAYS_INCLUDE, false);
 
 					// Handle search query.
-					if (query != null) {
+					if (query != null && !alwaysInclude) {
 						String attribute = row.getProperties().optString(filterAttribute, null);
 
 						if (attribute != null) {
@@ -745,8 +751,9 @@ public class TiTableView extends TiSwipeRefreshLayout implements OnSearchChangeL
 			}
 		}
 
+		Parcelable recyclerViewState = recyclerView.getLayoutManager().onSaveInstanceState();
 		// Notify adapter of changes on UI thread.
-		this.adapter.update(this.rows, force);
+		this.adapter.update(rows, force);
 
 		// FIXME: This is not an ideal workaround for an issue where recycled items that were in focus
 		//        lose their focus when the data set changes. There are improvements to be made here.
@@ -787,6 +794,7 @@ public class TiTableView extends TiSwipeRefreshLayout implements OnSearchChangeL
 						}
 					}
 				}
+				recyclerView.getLayoutManager().onRestoreInstanceState(recyclerViewState);
 			}
 		});
 	}
