@@ -53,10 +53,6 @@ import android.view.ViewAnimationUtils;
 	TiC.PROPERTY_BACKGROUND_SELECTED_IMAGE,
 	TiC.PROPERTY_BACKGROUND_FOCUSED_IMAGE,
 	TiC.PROPERTY_BACKGROUND_DISABLED_IMAGE,
-	TiC.PROPERTY_BACKGROUND_COLOR,
-	TiC.PROPERTY_BACKGROUND_SELECTED_COLOR,
-	TiC.PROPERTY_BACKGROUND_FOCUSED_COLOR,
-	TiC.PROPERTY_BACKGROUND_DISABLED_COLOR,
 	TiC.PROPERTY_BACKGROUND_PADDING,
 	TiC.PROPERTY_BACKGROUND_GRADIENT,
 	// border properties
@@ -854,11 +850,14 @@ public abstract class TiViewProxy extends KrollProxy
 	}
 
 	@Kroll.method
-	public void animate(Object arg, @Kroll.argument(optional = true) KrollFunction callback)
+	public void animate(@Kroll.argument(optional = true) Object arg,
+		@Kroll.argument(optional = true) KrollFunction callback)
 	{
 		synchronized (pendingAnimationLock)
 		{
-			if (arg instanceof HashMap) {
+			if (arg == null) {
+				stopAnimation();
+			} else if (arg instanceof HashMap) {
 				@SuppressWarnings("rawtypes")
 				HashMap options = (HashMap) arg;
 				pendingAnimation = new TiAnimationBuilder();
@@ -876,6 +875,16 @@ public abstract class TiViewProxy extends KrollProxy
 			}
 
 			handlePendingAnimation(false);
+		}
+	}
+
+	@Kroll.method
+	public void stopAnimation()
+	{
+		TiUIView tiv = peekView();
+
+		if (tiv != null) {
+			tiv.stopAnimation();
 		}
 	}
 
@@ -1055,6 +1064,12 @@ public abstract class TiViewProxy extends KrollProxy
 		return this.parent.get();
 	}
 
+	@Kroll.setProperty
+	public void setBackgroundColor(String color)
+	{
+		setPropertyAndFire(TiC.PROPERTY_BACKGROUND_COLOR, color);
+	}
+
 	@Kroll.getProperty
 	public String getBackgroundColor()
 	{
@@ -1087,6 +1102,12 @@ public abstract class TiViewProxy extends KrollProxy
 		return TiUIHelper.getBackgroundColorForState(tiBackgroundDrawable, TiUIHelper.BACKGROUND_DEFAULT_STATE_1);
 	}
 
+	@Kroll.setProperty
+	public void setBackgroundSelectedColor(String color)
+	{
+		setPropertyAndFire(TiC.PROPERTY_BACKGROUND_SELECTED_COLOR, color);
+	}
+
 	@Kroll.getProperty
 	public String getBackgroundSelectedColor()
 	{
@@ -1109,6 +1130,12 @@ public abstract class TiViewProxy extends KrollProxy
 		return TiUIHelper.getBackgroundColorForState(backgroundDrawable, TiUIHelper.BACKGROUND_SELECTED_STATE);
 	}
 
+	@Kroll.setProperty
+	public void setBackgroundFocusedColor(String color)
+	{
+		setPropertyAndFire(TiC.PROPERTY_BACKGROUND_FOCUSED_COLOR, color);
+	}
+
 	@Kroll.getProperty
 	public String getBackgroundFocusedColor()
 	{
@@ -1129,6 +1156,12 @@ public abstract class TiViewProxy extends KrollProxy
 			return null;
 		}
 		return TiUIHelper.getBackgroundColorForState(backgroundDrawable, TiUIHelper.BACKGROUND_FOCUSED_STATE);
+	}
+
+	@Kroll.setProperty
+	public void setBackgroundDisabledColor(String color)
+	{
+		setPropertyAndFire(TiC.PROPERTY_BACKGROUND_DISABLED_COLOR, color);
 	}
 
 	@Kroll.getProperty
