@@ -228,7 +228,7 @@
 {
   CNSaveRequest *saveRequest = [[CNSaveRequest alloc] init];
   [saveRequest deleteGroup:[[group mutableCopy] autorelease]];
-  // Do not be tempted to autorelease here. https://github.com/tidev/titanium_mobile/commit/a0d4a50d51f1afe85f92cf9e0d2ce8cca08fcf2f
+  // Do not be tempted to autorelease here. https://github.com/tidev/titanium-sdk/commit/a0d4a50d51f1afe85f92cf9e0d2ce8cca08fcf2f
   // It will be released in ContactsModule.m line 315 in (void)save
   return saveRequest;
 }
@@ -237,7 +237,7 @@
 {
   CNSaveRequest *saveRequest = [[CNSaveRequest alloc] init];
   [saveRequest addGroup:group toContainerWithIdentifier:containerIdentifier];
-  // Do not be tempted to autorelease here. https://github.com/tidev/titanium_mobile/commit/a0d4a50d51f1afe85f92cf9e0d2ce8cca08fcf2f
+  // Do not be tempted to autorelease here. https://github.com/tidev/titanium-sdk/commit/a0d4a50d51f1afe85f92cf9e0d2ce8cca08fcf2f
   // It will be released in ContactsModule.m line 315 in (void)save
   return saveRequest;
 }
